backend/bot/requirements.txt
python-telegram-bot
httpx
sqlalchemy
psycopg2-binary
boto3
python-jose[cryptography]
